+++ microweather/Models/Observation.cs
@@ -13,21 +13,21 @@ namespace microweather.Models
 
         public int altitude { get; set; }
 
-        public int temperature { get; set; }
+        public float temperature { get; set; }
 
         public int humidity { get; set; }
 
         public int pressure { get; set; }
 
-        public int wind_speed { get; set; }
+        public float wind_speed { get; set; }
 
         public string wind_direction { get; set; }
 
         public int soil_moisture { get; set; }
 
-        public int soil_temp { get; set; }
+        public float soil_temp { get; set; }
 
-        public int rain_rate { get; set; }
+        public float rain_rate { get; set; }
 
         public string image { get; set; }
